moved the HelloWorldResource to the client project and removed swagger annotations
