+++ integration/client/src/main/java/com/cerner/beadledom/integration/client/HelloWorldResource.java
@@ -1,11 +1,8 @@
-package com.cerner.beadledom.integration.api;
+package com.cerner.beadledom.integration.client;
 
 import com.cerner.beadledom.integration.api.model.HelloWorldDto;
 import com.cerner.beadledom.integration.api.model.OffsetPaginatedListHelloWorldDto;
 import com.cerner.beadledom.jaxrs.GenericResponse;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiResponse;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
@@ -16,7 +13,6 @@ import javax.ws.rs.core.MediaType;
  *
  * @author Nick Behrens
  */
-@Api(value = "/hello", description = "Retrieve hello world data")
 @Path("/hello")
 public interface HelloWorldResource {
 
@@ -25,10 +21,6 @@ public interface HelloWorldResource {
    *
    * @return an {@link OffsetPaginatedListHelloWorldDto}
    */
-  @ApiOperation(
-      value = "Retrieves hello world data.",
-      response = HelloWorldDto.class)
-  @ApiResponse(code = 200, message = "Success")
   @GET
   @Produces(MediaType.APPLICATION_JSON)
   GenericResponse<OffsetPaginatedListHelloWorldDto> getHelloWorld();
